Standardize initialization parameters
The previous version was just ugly. Now it is explicit that you are
given a lot of freedom with initialization but the configuration
property will be correctly populated with default values if you don't
supply them. There are now separate types for options and configuration

Also, interesting issue with tests. See the TODO comment in the
constructor in cube-timer.ts