Initial implementation (#1)
* [WIP] - add types

* Add tiny-timer typings

* [WIP] - Fleshing it out

* [WIP] - more fleshing out

* [WIP] - Main functionality is there, needs tests, cleanup and API delibaration

* Don't throw if stop is called on a stopped timer. log and return

* Start using eslint and never use that TS project starter again!

* Standardize initialization parameters

The previous version was just ugly. Now it is explicit that you are
given a lot of freedom with initialization but the configuration
property will be correctly populated with default values if you don't
supply them. There are now separate types for options and configuration

Also, interesting issue with tests. See the TODO comment in the
constructor in cube-timer.ts

* Add a time property and remove the unnecessary TimerStatus.Running enum value

* Add a simpler start() api. Not sure if this is the best idea